Move ping back into tasks module. Rely on job storage for job existence checks.
