+++ kolibri/core/analytics/utils.py
@@ -9,7 +9,6 @@ import sys
 import requests
 from dateutil import parser
 from django.core.serializers.json import DjangoJSONEncoder
-from django.db import connection
 from django.db import transaction
 from django.db.models import Count
 from django.db.models import Max
@@ -21,9 +20,6 @@ from django.utils.timezone import get_current_timezone
 from django.utils.timezone import localtime
 from le_utils.constants import content_kinds
 from morango.models import InstanceIDModel
-from requests.exceptions import ConnectionError
-from requests.exceptions import RequestException
-from requests.exceptions import Timeout
 
 import kolibri
 from .constants import nutrition_endpoints
@@ -45,9 +41,6 @@ from kolibri.core.logger.models import ContentSessionLog
 from kolibri.core.logger.models import ContentSummaryLog
 from kolibri.core.logger.models import MasteryLog
 from kolibri.core.logger.models import UserSessionLog
-from kolibri.core.tasks.decorators import register_task
-from kolibri.core.tasks.main import job_storage
-from kolibri.core.tasks.utils import get_current_job
 from kolibri.core.utils.lock import db_lock
 from kolibri.utils import conf
 from kolibri.utils.server import installation_type
@@ -55,9 +48,7 @@ from kolibri.utils.time_utils import local_now
 
 logger = logging.getLogger(__name__)
 
-DEFAULT_PING_INTERVAL = 24 * 60
-DEFAULT_PING_CHECKRATE = 15
-DEFAULT_PING_JOB_ID = "0"
+
 DEFAULT_SERVER_URL = "https://telemetry.learningequality.org"
 
 USER_THRESHOLD = 10
@@ -488,51 +479,3 @@ def ping_once(started, server=DEFAULT_SERVER_URL):
     if "id" in data:
         stat_data = perform_statistics(server, data["id"])
         create_and_update_notifications(stat_data, nutrition_endpoints.STATISTICS)
-
-
-@register_task(job_id=DEFAULT_PING_JOB_ID)
-def _ping(started, server, checkrate):
-    try:
-        ping_once(started, server=server)
-        connection.close()
-        return
-    except ConnectionError:
-        logger.warning(
-            "Ping failed (could not connect). Trying again in {} minutes.".format(
-                checkrate
-            )
-        )
-    except Timeout:
-        logger.warning(
-            "Ping failed (connection timed out). Trying again in {} minutes.".format(
-                checkrate
-            )
-        )
-    except RequestException as e:
-        logger.warning(
-            "Ping failed ({})! Trying again in {} minutes.".format(e, checkrate)
-        )
-    connection.close()
-    job = get_current_job()
-    if job and job in job_storage:
-        job_storage.change_execution_time(
-            job, local_now() + datetime.timedelta(seconds=checkrate * 60)
-        )
-
-
-def schedule_ping(
-    server=DEFAULT_SERVER_URL,
-    checkrate=DEFAULT_PING_CHECKRATE,
-    interval=DEFAULT_PING_INTERVAL,
-):
-    # If pinging is not disabled by the environment
-    if not conf.OPTIONS["Deployment"]["DISABLE_PING"]:
-        # Scheduler needs datetime object, but job needs (serializable) string
-        now = local_now()
-        started = now.isoformat()
-        _ping.enqueue_at(
-            now,
-            interval=interval * 60,
-            repeat=None,
-            kwargs=dict(started=started, server=server, checkrate=checkrate),
-        )
